AdventOfCode.Tests/Days/Day19Tests.cs
using System;
using AdventOfCode.Days;
using FluentAssertions;
using Xunit;

namespace AdventOfCode.Tests.Days
{
    
    public class Day19Tests
    {
        private readonly ISolution _sut = new Day19();
        
        [Fact]
        public void PartOne_WhenCalled_DoesNotThrowNotImplementedException()
        {
            Action act = () =>  _sut.PartOne(_sut.Input());
            
            act.Should().NotThrow<NotImplementedException>();
        }
        
        [Fact]
        public void PartTwo_WhenCalled_DoesNotThrowNotImplementedException()
        {
            Action act = () =>  _sut.PartTwo(_sut.Input());
            
            act.Should().NotThrow<NotImplementedException>();
        }
        
        
        [Fact]
        public void PartTwo_WhenCalled_RunsToCompletion()
        {
            var res =  _sut.PartTwo(_sut.Input());
            
            res.Should().Be("207");
        }
    }
}